Files table working with axios, configurations api routes fixed

+++ routes/web.php
@@ -102,56 +102,56 @@ Route::prefix('api')->group(function () {
 
 // Academic periods: Periodos académicos
 Route::prefix('api')->group(function () {
-    Route::get('/periodos-academicos', [AcademicPeriodController::class, 'apiIndex']);
-    Route::post('/periodos-academicos', [AcademicPeriodController::class, 'store']);
-    Route::put('/periodos-academicos/{period}', [AcademicPeriodController::class, 'update']);
-    Route::delete('/periodos-academicos/{period}', [AcademicPeriodController::class, 'destroy']);
+    Route::get('/periodos-academicos', [AcademicPeriodController::class, 'apiIndex'])->name('api.academicPeriods.index');
+    Route::post('/periodos-academicos', [AcademicPeriodController::class, 'store'])->name('api.academicPeriods.store');
+    Route::put('/periodos-academicos/{period}', [AcademicPeriodController::class, 'update'])->name('api.academicPeriods.update');
+    Route::delete('/periodos-academicos/{period}', [AcademicPeriodController::class, 'destroy'])->name('api.academicPeriods.destroy');
 });
 Route::get('/configuracion/periodos-academicos', [AcademicPeriodController::class, 'index'])->name('academicPeriods.index');
 
 // Courses: Carreras
 Route::prefix('api')->group(function () {
-    Route::get('/carreras', [CourseController::class, 'apiIndex']);
-    Route::post('/carreras', [CourseController::class, 'store']);
-    Route::put('/carreras/{course}', [CourseController::class, 'update']);
-    Route::delete('/carreras/{course}', [CourseController::class, 'destroy']);
+    Route::get('/carreras', [CourseController::class, 'apiIndex'])->name('api.courses.index');
+    Route::post('/carreras', [CourseController::class, 'store'])->name('api.courses.store');
+    Route::put('/carreras/{course}', [CourseController::class, 'update'])->name('api.courses.update');
+    Route::delete('/carreras/{course}', [CourseController::class, 'destroy'])->name('api.courses.destroy');
 });
 Route::get('/configuracion/carreras', [CourseController::class, 'index'])->name('courses.index');
 
 // Thesis areas: Áreas de titulación
 Route::prefix('api')->group(function () {
-    Route::get('/areas-titulacion', [ThesisAreaController::class, 'apiIndex']);
-    Route::post('/areas-titulacion', [ThesisAreaController::class, 'store']);
-    Route::put('/areas-titulacion/{area}', [ThesisAreaController::class, 'update']);
-    Route::delete('/areas-titulacion/{area}', [ThesisAreaController::class, 'destroy']);
+    Route::get('/areas-titulacion', [ThesisAreaController::class, 'apiIndex'])->name('api.thesisAreas.index');
+    Route::post('/areas-titulacion', [ThesisAreaController::class, 'store'])->name('api.thesisAreas.store');
+    Route::put('/areas-titulacion/{area}', [ThesisAreaController::class, 'update'])->name('api.thesisAreas.update');
+    Route::delete('/areas-titulacion/{area}', [ThesisAreaController::class, 'destroy'])->name('api.thesisAreas.destroy');
 });
 Route::get('/configuracion/areas-titulacion', [ThesisAreaController::class, 'index'])->name('thesisAreas.index');
 
 // Roles
 Route::prefix('api')->group(function () {
-    Route::get('/roles', [RoleController::class, 'apiIndex']);
-    Route::post('/roles', [RoleController::class, 'store']);
-    Route::put('/roles/{role}', [RoleController::class, 'update']);
-    Route::delete('/roles/{role}', [RoleController::class, 'destroy']);
+    Route::get('/roles', [RoleController::class, 'apiIndex'])->name('api.roles.index');
+    Route::post('/roles', [RoleController::class, 'store'])->name('api.roles.store');
+    Route::put('/roles/{role}', [RoleController::class, 'update'])->name('api.roles.update');
+    Route::delete('/roles/{role}', [RoleController::class, 'destroy'])->name('api.roles.destroy');
 });
 Route::get('/configuracion/roles', [RoleController::class, 'index'])->name('roles.index');
 
 // Permissions: Permisos
 //TODO: check if permissions are needed at all
 Route::prefix('api')->group(function () {
-    Route::get('/permisos', [PermissionController::class, 'apiIndex']);
-    Route::post('/permisos', [PermissionController::class, 'store']);
-    Route::put('/permisos/{permission}', [PermissionController::class, 'update']);
-    Route::delete('/permisos/{permission}', [PermissionController::class, 'destroy']);
+    Route::get('/permisos', [PermissionController::class, 'apiIndex'])->name('api.permissions.index');
+    Route::post('/permisos', [PermissionController::class, 'store'])->name('api.permissions.store');
+    Route::put('/permisos/{permission}', [PermissionController::class, 'update'])->name('api.permissions.update');
+    Route::delete('/permisos/{permission}', [PermissionController::class, 'destroy'])->name('api.permissions.destroy');
 });
 Route::get('/configuracion/permisos', [PermissionController::class, 'index'])->name('permissions.index');
 
 // Users: Usuarios
 Route::prefix('api')->group(function () {
-    Route::get('/usuarios', [UserController::class, 'apiIndex']);
-    Route::post('/usuarios', [UserController::class, 'store']);
-    Route::put('/usuarios/{user}', [UserController::class, 'update']);
-    Route::delete('/usuarios/{user}', [UserController::class, 'destroy']);
+    Route::get('/usuarios', [UserController::class, 'apiIndex'])->name('api.users.index');
+    Route::post('/usuarios', [UserController::class, 'store'])->name('api.users.store');
+    Route::put('/usuarios/{user}', [UserController::class, 'update'])->name('api.users.update');
+    Route::delete('/usuarios/{user}', [UserController::class, 'destroy'])->name('api.users.destroy');
 });
 Route::get('/configuracion/usuarios', [UserController::class, 'index'])->name('users.index');
 
@@ -176,6 +176,7 @@ Route::put('/titulacion/estudiantes/proceso/actualizar/{graduation}', [Graduatio
 Route::get('/titulacion/periodos-academicos', [GraduationController::class, 'periods'])->name('graduation.periods');
 Route::get('/titulacion/periodos-academicos/{period}/estudiantes', [GraduationController::class, 'studentsInPeriod'])->name('graduation.studentsInPeriod');
 Route::get('/titulacion/estudiantes/{student}/documentos', [GraduationFilesController::class, 'studentFiles'])->name('graduationFiles.studentFiles');
+Route::get('/api/graduationFiles/{student}', [GraduationFilesController::class, 'apiStudentFiles'])->name('api.graduationFiles.studentFiles');
 
 
 // Reviewers: Evaluadores
@@ -183,22 +184,22 @@ Route::get('/titulacion/evaluadores', [GraduationController::class, 'reviewers']
 Route::get('/titulacion/evaluadores/docente/{professor}', [GraduationController::class, 'processesByProfessor'])->name('graduation.processesByProfessor');
 Route::prefix('api/titulacion/evaluadores')->group(function () {
     Route::get('/get-reviewers-by-students', [GraduationController::class, 'getReviewersByStudents'])
-        ->name('graduation.getReviewersByStudents');
+        ->name('api.graduation.getReviewersByStudents');
     Route::get('/get-reviewers-by-professors', [GraduationController::class, 'getReviewersByProfessors'])
-        ->name('graduation.getReviewersByProfessors');
+        ->name('api-graduation.getReviewersByProfessors');
     Route::get('/reviewers/processes/professor/{professor}', [GraduationController::class, 'processesByProfessor'])->name('graduation.processesByProfessor');
     Route::get('/get-processes-as-advisor/{id}', [GraduationController::class, 'getProcessesAsAdvisor'])
-        ->name('graduation.getProcessesAsAdvisor');
+        ->name('api.graduation.getProcessesAsAdvisor');
     Route::get('/get-processes-as-reader/{id}', [GraduationController::class, 'getProcessesAsReader'])
-        ->name('graduation.getProcessesAsReader');
+        ->name('api.graduation.getProcessesAsReader');
 });
 
 // Graduation statistics: Reportes de titulación
 Route::get('/titulacion/reportes', [GraduationController::class, 'statistics'])->name('graduation.statistics');
 Route::prefix('api/titulacion/reportes')->group(function () {
-    Route::get('/planes-por-caducar', [GraduationController::class, 'getPlansDueToExpire']);
-    Route::get('/numero-de-matricula', [GraduationController::class, 'getRegistrationTimes']);
-    Route::get('/graduados-por-fecha/{start}/{end}', [GraduationController::class, 'getGraduatesByDate']);
+    Route::get('/planes-por-caducar', [GraduationController::class, 'getPlansDueToExpire'])->name('api.graduation.plansDueToExpire');
+    Route::get('/numero-de-matricula', [GraduationController::class, 'getRegistrationTimes'])->name('api.graduation.registrationTimes');
+    Route::get('/graduados-por-fecha/{start}/{end}', [GraduationController::class, 'getGraduatesByDate'])->name('api.graduation.graduatesByDate');
 });
